Include response body on non-101 status
When the server rejects the connection, it can supply a message,
which becomes the response body.

However, the client did not expose this manage in any form until now.

This commit introduces a special exception to be thrown on non-101
response, ResponseStatusError, and this exception includes the full
response body.